% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_analysis.R
\name{mlr_analysis}
\alias{mlr_analysis}
\title{Model analysis}
\usage{
mlr_analysis(formula, dataset)
}
\arguments{
\item{formula}{a formula in form of response ~ predictor1 + predictor2 + ...}

\item{data}{a data frame containing all needed variables in the formula.}
}
\value{
the summary and anova table which contains outcomes of mlr model.
}
\description{
This function calculates and analysis multiple linear regression
}
\examples{
mlr_analysis(formula = mpg ~ hp + cyl, data = mtcars)
mlr_analysis(formula = mpg ~ hp + cyl + wt, data = mtcars)
mlr_analysis(formula = mpg ~ hp + cyl + wt + qsec, data = mtcars)
mlr_analysis(formula = mpg ~ hp + cyl + wt + qsec + cyl * wt, data = mtcars)

}
