% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_analysis.R
\name{mlr_analysis}
\alias{mlr_analysis}
\title{Multiple linear model analysis}
\usage{
mlr_analysis(formula, input_data)
}
\arguments{
\item{formula}{a formula in form of response ~ predictor1 + predictor2 + ...}

\item{input_data}{a data frame containing all needed variables in the formula.}
}
\value{
the summary table which contains residuals, fitted_values, df,
coeff_table（which include Estimates, Std.Error, t value and p value,
R_squared, adj_R_squared, F_statistic, p-value of F_test and CI_95 of
the input multiple linear model.
}
\description{
This function calculates and analysis multiple linear regression
}
\examples{
mlr_analysis(formula = mpg ~ hp + cyl, data = mtcars)
mlr_analysis(formula = mpg ~ hp + cyl + wt, data = mtcars)
mlr_analysis(formula = mpg ~ hp + cyl + wt + qsec, data = mtcars)
mlr_analysis(formula = mpg ~ hp + cyl + wt + qsec + cyl * wt, data = mtcars)

}
